feat: 体力气力系统完整实现 + 关系成长可配置化 + 指令总线扩展
体力/气力系统:
- VitalService: DOWNBASE扣减、疲劳计算(CULC_TIRED)、自然恢复/睡眠/小憩/泡澡/通用恢复
- VitalGate: 体力归零封所有指令, 气力归零封消耗气力指令
- 体力归零晕倒: 自动睡眠恢复 + 推进时间到次日黎明 + ActionResult.fainted标志
- MAXBASE上限强制执行, fatigue加入CharacterState及存档序列化
- 新增recovery类指令: nap(小憩)/sleep(就寝)/bathe(泡澡)
- 旧nap改为relax_together(一起放松), 加required_stage=friendly

关系成长可配置化:
- relationship_growth.toml + relationship_stages.toml 动态读取
- 可配置的情感/信赖增长公式(favor_calc/trust_calc)
- 测试辅助stages.py + expected.py工厂函数

指令总线扩展:
- settlement_rules扩展CUP/CDOWN、CFLAG直接写入
- source_extra(TALENT效果乘算/加算)接入结算管道
- CommandService委托VitalService处理DOWNBASE和恢复操作
- GameLoop.advance_time()接入自然恢复

新增35个vital系统测试, 234总测试全部通过

+++ src/eral/content/settlement.py
@@ -17,7 +17,7 @@ class SettlementRule:
     target_family: AxisFamily
     target_key: str | None = None
     target_index: int | None = None
-    scale: int = 1
+    scale: float = 1.0
 
 
 def load_settlement_rules(path: Path) -> tuple[SettlementRule, ...]:
@@ -32,7 +32,7 @@ def load_settlement_rules(path: Path) -> tuple[SettlementRule, ...]:
             target_family=AxisFamily(item["target_family"]),
             target_key=item.get("target_key"),
             target_index=item.get("target_index"),
-            scale=int(item.get("scale", 1)),
+            scale=float(item.get("scale", 1)),
         )
         for item in raw_data.get("rules", [])
     )
